+++ engine/flink/model-util/src/main/java/pl/touk/nussknacker/engine/flink/util/transformer/join/BranchType.java
@@ -1,0 +1,6 @@
+package pl.touk.nussknacker.engine.flink.util.transformer.join;
+
+public enum BranchType {
+    MAIN, JOINED;
+}
+