+++ engine/flink/model-util/src/main/java/pl/touk/nussknacker/engine/flink/util/transformer/outer/BranchType.java
@@ -1,6 +1,0 @@
-package pl.touk.nussknacker.engine.flink.util.transformer.outer;
-
-public enum BranchType {
-    MAIN, JOINED;
-}
-