+++ engine/interpreter/src/test/scala/pl/touk/nussknacker/engine/compile/ProcessValidatorSpec.scala
@@ -93,10 +93,50 @@ class ProcessValidatorSpec extends FunSuite with Matchers with Inside {
     ObjectDefinition.noParam,
     ExpressionDefinition(
       Map("processHelper" -> ObjectDefinition(List(), Typed(ProcessHelper.getClass), List("cat1"), SingleNodeConfig.zero)),
-      List.empty, List.empty, LanguageConfiguration.default, optimizeCompilation = false, strictTypeChecking = true, dictionaries = Map.empty, hideMetaVariable = false, strictMethodsChecking = true
+      List.empty, List.empty, LanguageConfiguration.default, optimizeCompilation = false, strictTypeChecking = true, dictionaries = Map.empty, hideMetaVariable = false, strictMethodsChecking = true, staticMethodInvocationsChecking = true
     ),
     ClassExtractionSettings.Default
   )
+
+  test("Validation of Type Reference using accessible class, success scenario") {
+
+    val typeReferenceWithValidClass = "T(String).copyValueOf()"
+
+    val testProcess =
+      EspProcessBuilder
+        .id("TypeReferenceClassValidationSuccess")
+        .exceptionHandler()
+        .source("source1", "source")
+        .filter("filter1", typeReferenceWithValidClass)
+        .sink("id1", "#input", "sink")
+
+    val compilationResult = validate(testProcess, baseDefinition)
+
+    compilationResult.result should matchPattern {
+      case Valid(_) =>
+    }
+
+  }
+
+  test("Validation of Type Reference using inaccessible class, failure scenario") {
+
+    val typeReferenceWithValidClass = "T(System).exit()"
+
+    val testProcess =
+      EspProcessBuilder
+        .id("TypeReferenceClassValidationFailure")
+        .exceptionHandler()
+        .source("source1", "source")
+        .filter("filter1", typeReferenceWithValidClass)
+        .sink("id1", "#input", "sink")
+
+    val compilationResult = validate(testProcess, baseDefinition)
+
+    compilationResult.result should matchPattern {
+      case Invalid(NonEmptyList(ExpressionParseError(_, _, _, _), _)) =>
+    }
+
+  }
 
   test("validated with success") {
     val correctProcess = EspProcessBuilder